rockfile: Add an update-crc command
This can be used, for example, to make reproducible builds from the
output of `${rkbin}/tools/boot_merger`.

Using this savage invocation of `dd` on the binary file `boot_merger`
produces removes the timestamp.

```
dd if=/dev/zero of="${out}" bs=1 seek="$((0x0E))" count=7 conv=notrunc
```

(Hint: Using `faketime` won't work since this is a static binary!)

Once this is done, `rkdeveloptool` will not want to send the file over
USB, as it has an invalid CRC.

> ```
> Opening loader failed, exiting download boot!
> ```

Using the `update-crc`, the CRC can be fixed, and `rkdeveloptool` is
happy with the file.

Currently, rockchiprs' `rockusb` commands don't validate the CRC. This
shows the work necessary to validate the checksum, if desired.

My next step is to add a command to null-out the timestamp of a
RkBootHeader, such that using `rockfile erase-timestamp` will make it
easy to produce reproducible outputs from the vendor's `boot_merger`.

Eventually, producing those files from within rockchiprs would be
better, but one step at a time.